+++ app_utils/allianceauth.py
@@ -1,6 +1,7 @@
 from django.contrib.auth.models import User, Permission
 from django.db.models import Q
 
+from allianceauth.views import NightModeRedirectView
 from allianceauth.notifications import notify
 
 
@@ -18,3 +19,8 @@ def notify_admins(message: str, title: str, level="info") -> None:
             notify(user, title=title, message=message, level=level)
     except Permission.DoesNotExist:
         pass
+
+
+def is_night_mode(request) -> bool:
+    """Returns True if the current user session is in night mode, else False"""
+    return NightModeRedirectView.night_mode_state(request)